feat: add ROICalculator to calculate return on investment

+++ src/main/java/lotto/view/OutputView.java
@@ -31,7 +31,7 @@ public class OutputView {
         });
     }
 
-    public static void printReturnOnInvestment(int returnOnInvestment) {
+    public static void printReturnOnInvestment(double returnOnInvestment) {
         System.out.println("총 수익률은 " + returnOnInvestment + "입니다.(기준이 1이기 때문에 결과적으로 손해라는 의미임)");
     }
